Revert "Store/update geneticengine"
This reverts commit 5f99c636832fa1f6cd2da2f10b1a6e69994ab46c.
