+++ official_competitors/geneticengine/environment.lock.yml
@@ -3,8 +3,8 @@ channels:
   - conda-forge
 dependencies:
   - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - alsa-lib=1.2.3.2=h166bdaf_0
+  - _openmp_mutex=4.5=1_gnu
+  - alsa-lib=1.2.3=h516909a_0
   - argon2-cffi=21.3.0=pyhd8ed1ab_0
   - argon2-cffi-bindings=21.2.0=py39hb9d737c_2
   - asttokens=2.0.5=pyhd8ed1ab_0
@@ -36,30 +36,30 @@ dependencies:
   - fontconfig=2.14.0=h8e229c2_0
   - fonts-conda-ecosystem=1=0
   - fonts-conda-forge=1=0
-  - fonttools=4.33.3=py39hb9d737c_0
+  - fonttools=4.32.0=py39hb9d737c_0
   - freetype=2.10.4=h0708190_1
   - gettext=0.19.8.1=h73d1719_1008
   - giflib=5.2.1=h36c2ea0_2
   - gmp=6.2.1=h58526e2_0
   - gmpy2=2.1.2=py39h78fa15d_0
-  - gst-plugins-base=1.20.2=hcf0ee16_0
-  - gstreamer=1.20.2=hd4edc92_0
+  - gst-plugins-base=1.20.1=hcf0ee16_1
+  - gstreamer=1.20.1=hd4edc92_1
   - icu=69.1=h9c3ff4c_0
   - importlib-metadata=4.11.3=py39hf3d152e_1
-  - importlib_resources=5.7.1=pyhd8ed1ab_0
+  - importlib_resources=5.7.0=pyhd8ed1ab_0
   - iniconfig=1.1.1=pyh9f0ad1d_0
   - ipykernel=6.13.0=py39hef51801_0
-  - ipython=8.3.0=py39hf3d152e_0
+  - ipython=8.2.0=py39hf3d152e_0
   - ipython_genutils=0.2.0=py_1
   - ipywidgets=7.7.0=pyhd8ed1ab_0
   - jbig=2.1=h7f98852_2003
   - jedi=0.18.1=py39hf3d152e_1
-  - jinja2=3.1.2=pyhd8ed1ab_0
+  - jinja2=3.1.1=pyhd8ed1ab_0
   - joblib=1.1.0=pyhd8ed1ab_0
-  - jpeg=9e=h166bdaf_1
-  - jsonschema=4.5.1=pyhd8ed1ab_0
+  - jpeg=9e=h7f98852_0
+  - jsonschema=4.4.0=pyhd8ed1ab_0
   - jupyter=1.0.0=py39hf3d152e_7
-  - jupyter_client=7.3.1=pyhd8ed1ab_0
+  - jupyter_client=7.2.2=pyhd8ed1ab_1
   - jupyter_console=6.4.3=pyhd8ed1ab_0
   - jupyter_core=4.9.2=py39hf3d152e_0
   - jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
@@ -80,11 +80,11 @@ dependencies:
   - libedit=3.1.20191231=he28a2e2_2
   - libevent=2.1.10=h28343ad_4
   - libffi=3.4.2=h7f98852_5
-  - libgcc-ng=11.2.0=h1d223b6_16
-  - libgfortran-ng=11.2.0=h69a702a_16
-  - libgfortran5=11.2.0=h5c6108e_16
+  - libgcc-ng=11.2.0=h1d223b6_15
+  - libgfortran-ng=11.2.0=h69a702a_15
+  - libgfortran5=11.2.0=h5c6108e_15
   - libglib=2.70.2=h174f98d_4
-  - libgomp=11.2.0=h1d223b6_16
+  - libgomp=11.2.0=h1d223b6_15
   - libiconv=1.16=h516909a_0
   - liblapack=3.9.0=14_linux64_openblas
   - libllvm13=13.0.1=hf817b99_2
@@ -95,7 +95,7 @@ dependencies:
   - libpng=1.6.37=h21135ba_2
   - libpq=14.2=h676c864_0
   - libsodium=1.0.18=h36c2ea0_1
-  - libstdcxx-ng=11.2.0=he4da1e4_16
+  - libstdcxx-ng=11.2.0=he4da1e4_15
   - libtiff=4.3.0=h542a066_3
   - libuuid=2.32.1=h7f98852_1000
   - libvorbis=1.3.7=h9c3ff4c_0
@@ -107,29 +107,29 @@ dependencies:
   - libzlib=1.2.11=h166bdaf_1014
   - lz4-c=1.9.3=h9c3ff4c_1
   - markupsafe=2.1.1=py39hb9d737c_1
-  - matplotlib=3.5.2=py39hf3d152e_0
-  - matplotlib-base=3.5.2=py39h700656a_0
+  - matplotlib=3.5.1=py39hf3d152e_0
+  - matplotlib-base=3.5.1=py39h2fa2bec_0
   - matplotlib-inline=0.1.3=pyhd8ed1ab_0
   - mistune=0.8.4=py39h3811e60_1005
   - mpc=1.2.1=h9f54685_0
   - mpfr=4.1.0=h9202a9a_1
   - mpmath=1.2.1=pyhd8ed1ab_0
   - munkres=1.1.4=pyh9f0ad1d_0
-  - mysql-common=8.0.29=h26416b9_0
-  - mysql-libs=8.0.29=hbc51c84_0
-  - nbclient=0.6.3=pyhd8ed1ab_0
+  - mysql-common=8.0.28=h26416b9_3
+  - mysql-libs=8.0.28=hbc51c84_3
+  - nbclient=0.6.0=pyhd8ed1ab_0
   - nbconvert=6.5.0=pyhd8ed1ab_0
   - nbconvert-core=6.5.0=pyhd8ed1ab_0
   - nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
-  - nbformat=5.4.0=pyhd8ed1ab_0
+  - nbformat=5.3.0=pyhd8ed1ab_0
   - ncurses=6.3=h27087fc_1
   - nest-asyncio=1.5.5=pyhd8ed1ab_0
-  - notebook=6.4.11=pyha770c72_0
+  - notebook=6.4.10=pyha770c72_0
   - nspr=4.32=h9c3ff4c_1
   - nss=3.77=h2350873_0
-  - numpy=1.22.3=py39hc58783e_2
+  - numpy=1.22.3=py39h18676bf_2
   - openjpeg=2.4.0=hb52868f_1
-  - openssl=3.0.3=h166bdaf_0
+  - openssl=3.0.2=h166bdaf_1
   - packaging=21.3=pyhd8ed1ab_0
   - pandas=1.4.2=py39h1832856_1
   - pandoc=2.18=ha770c72_0
@@ -150,7 +150,7 @@ dependencies:
   - pure_eval=0.2.2=pyhd8ed1ab_0
   - py=1.11.0=pyh6c4a22f_0
   - pycparser=2.21=pyhd8ed1ab_0
-  - pygments=2.12.0=pyhd8ed1ab_0
+  - pygments=2.11.2=pyhd8ed1ab_0
   - pyparsing=3.0.8=pyhd8ed1ab_0
   - pyqt=5.12.3=py39hf3d152e_8
   - pyqt-impl=5.12.3=py39hde8b62d_8
@@ -158,7 +158,7 @@ dependencies:
   - pyqtchart=5.12=py39h0fcd23e_8
   - pyqtwebengine=5.12.1=py39h0fcd23e_8
   - pyrsistent=0.18.1=py39hb9d737c_1
-  - pytest=7.1.2=py39hf3d152e_0
+  - pytest=7.1.1=py39hf3d152e_1
   - python=3.9.12=h2660328_1_cpython
   - python-dateutil=2.8.2=pyhd8ed1ab_0
   - python-fastjsonschema=2.15.3=pyhd8ed1ab_0
@@ -168,7 +168,7 @@ dependencies:
   - qt=5.12.9=h1304e3e_6
   - qtconsole=5.3.0=pyhd8ed1ab_0
   - qtconsole-base=5.3.0=pyhd8ed1ab_0
-  - qtpy=2.1.0=pyhd8ed1ab_0
+  - qtpy=2.0.1=pyhd8ed1ab_0
   - readline=8.1=h46c0cb4_0
   - scikit-learn=0.24.1=py39h4dfa638_0
   - scipy=1.8.0=py39hee8e79c_1
@@ -176,7 +176,7 @@ dependencies:
   - setuptools=62.1.0=py39hf3d152e_0
   - six=1.16.0=pyh6c4a22f_0
   - soupsieve=2.3.1=pyhd8ed1ab_0
-  - sqlite=3.38.5=h4ff8645_0
+  - sqlite=3.38.2=h4ff8645_0
   - stack_data=0.2.0=pyhd8ed1ab_0
   - sympy=1.10.1=py39hf3d152e_0
   - terminado=0.13.3=py39hf3d152e_1
@@ -201,7 +201,7 @@ dependencies:
   - zstd=1.5.2=ha95c52a_0
   - pip:
     - charset-normalizer==2.0.12
-    - geneticengine==0.0.15
+    - geneticengine==0.0.14
     - idna==3.3
     - lark==1.1.2
     - pmlb==1.0.1.post3
