+++ submission/geneticengine/environment.yml
@@ -8,6 +8,7 @@ dependencies:
   - setuptools=62.1.0
   - pip=22.0.4
   - pip:
-    - geneticengine==0.0.14
+    - geneticengine==0.0.15
     - lark==1.1.2
     - z3-solver==4.8.15.0
+
