+++ submission/geneticengine/install.sh
@@ -1,0 +1,1 @@
+pip install geneticengine==0.0.15
